# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Marketing Simulation Lab API",
    "description": "API documentation for Digital Marketing Simulation Lab backend",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/billing/plans": {
      "post": {
        "tags": [
          "Admin Billing"
        ],
        "description": "Provisions a new billing/pricing plan (Admin only)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "code",
                  "priceMonthly",
                  "priceYearly",
                  "simulationLimit",
                  "studentLimit",
                  "instructorLimit",
                  "certificateLimit",
                  "reportExportLimit"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  },
                  "priceMonthly": {
                    "type": "number"
                  },
                  "priceYearly": {
                    "type": "number"
                  },
                  "currency": {
                    "type": "string",
                    "default": "INR"
                  },
                  "simulationLimit": {
                    "type": "integer"
                  },
                  "studentLimit": {
                    "type": "integer"
                  },
                  "instructorLimit": {
                    "type": "integer"
                  },
                  "certificateLimit": {
                    "type": "integer"
                  },
                  "reportExportLimit": {
                    "type": "integer"
                  },
                  "storageLimitMb": {
                    "type": "integer",
                    "default": 50
                  },
                  "features": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  },
                  "durationDays": {
                    "type": "integer",
                    "default": 30
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/billing/invoices/{id}/address": {
      "put": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ]
}